Break tainttoleration Filter plugins dependency on predicates package
